clean client.js
