+++ src/client.js
@@ -1,123 +1,22 @@
 import React from 'react'
 import { hydrate } from 'react-dom'
-import { BrowserRouter } from 'react-router-dom'
-import gql from 'graphql-tag'
-import { ApolloClient } from 'apollo-client'
-import { HttpLink } from 'apollo-link-http'
-import { ApolloLink } from 'apollo-link'
-import { InMemoryCache } from 'apollo-cache-inmemory'
-import { withClientState } from 'apollo-link-state'
+import { BrowserRouter as Router } from 'react-router-dom'
 import { ApolloProvider } from 'react-apollo'
-import App from './App'
 import { ThemeProvider } from 'styled-components'
 
-import './Utils/global-styles'
-import './Utils/icons'
+import 'isomorphic-fetch'
+
+import App from './App'
 import theme from './Utils/theme'
-import { WATCHED_KEY, FAV_KEY, getStorage } from './Utils/state'
-
-const defaultState = {
-    favorites: getStorage(FAV_KEY),
-    watched: getStorage(WATCHED_KEY),
-    hideViewed: false,
-    search: ''
-}
-
-const cache = new InMemoryCache()
-// .restore(window.__APOLLO_STATE__)
-
-export const stateLink = withClientState({
-    defaults: defaultState,
-    cache,
-    resolvers: {
-        Mutation: {
-            addFavorite: (_, { id }, { cache }) => {
-                const query = gql`
-                    query GetFavorites {
-                        favorites @client
-                    }
-                `
-
-                const previous = cache.readQuery({ query })
-                const data = {
-                    favorites: [...previous.favorites, id]
-                }
-
-                localStorage.setItem(FAV_KEY, JSON.stringify(data.favorites))
-
-                cache.writeQuery({ query, data })
-            },
-            removeFavorite: (_, { id }, { cache }) => {
-                const query = gql`
-                    query GetFavorites {
-                        favorites @client
-                    }
-                `
-
-                const previous = cache.readQuery({ query })
-                const data = {
-                    favorites: previous.favorites.filter(a => a !== id)
-                }
-
-                localStorage.setItem(FAV_KEY, JSON.stringify(data.favorites))
-
-                cache.writeQuery({ query, data })
-            },
-            addWatched: (_, { id }, { cache }) => {
-                const query = gql`
-                    query GetWatched {
-                        watched @client
-                    }
-                `
-
-                const previous = cache.readQuery({ query })
-                const data = {
-                    watched: [...previous.watched, id]
-                }
-
-                localStorage.setItem(WATCHED_KEY, JSON.stringify(data.watched))
-
-                cache.writeQuery({ query, data })
-            },
-            removeWatched: (_, { id }, { cache }) => {
-                const query = gql`
-                    query GetWatched {
-                        watched @client
-                    }
-                `
-
-                const previous = cache.readQuery({ query })
-                const data = {
-                    watched: previous.watched.filter(a => a !== id)
-                }
-
-                localStorage.setItem(WATCHED_KEY, JSON.stringify(data.watched))
-
-                cache.writeQuery({ query, data })
-            }
-        }
-    }
-})
-
-const client = new ApolloClient({
-    connectToDevTools: process.browser,
-    ssrMode: !process.browser,
-    link: ApolloLink.from([
-        stateLink,
-        new HttpLink({
-            uri: 'https://api.graphcms.com/simple/v1/cjhdcwrb98if90109o4pzawaq'
-        })
-    ]),
-    cache
-})
+import client from './Utils/stateLink'
 
 hydrate(
     <ApolloProvider client={client}>
-        <BrowserRouter>
-            <ThemeProvider theme={theme}>
+        <ThemeProvider theme={theme}>
+            <Router>
                 <App />
-            </ThemeProvider>
-        </BrowserRouter>
+            </Router>
+        </ThemeProvider>
     </ApolloProvider>,
     document.getElementById('root')
 )
@@ -125,3 +24,5 @@ hydrate(
 if (module.hot) {
     module.hot.accept()
 }
+
+// require('offline-plugin/runtime').install()
